# lib/api-zod/src/generated/types/registerUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Monjez - Smart Labor Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterUserBodyRole } from "./registerUserBodyRole";

export interface RegisterUserBody {
  name: string;
  phone: string;
  password: string;
  role: RegisterUserBodyRole;
  city?: string | null;
  bio?: string | null;
  israeliId?: string | null;
  businessId?: string | null;
}