# lib/api-zod/src/generated/types/sendOtpResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Monjez - Smart Labor Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface SendOtpResponse {
  message: string;
  /** OTP code (only returned in dev mode - remove in production) */
  otpCode?: string;
  cooldownSeconds: number;
}