# lib/api-zod/src/generated/types/updateMeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Monjez - Smart Labor Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMeBody {
  name?: string;
  bio?: string | null;
  city?: string | null;
  israeliId?: string | null;
  businessId?: string | null;
}