# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Monjez - Smart Labor Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  phone: string;
  role: UserRole;
  bio?: string | null;
  city?: string | null;
  israeliId?: string | null;
  businessId?: string | null;
  phoneVerified: boolean;
  trustScore?: number | null;
  creditBalance: number;
  createdAt: Date;
}